requirements.txt
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
black==21.9b0
click==8.0.1
colorama==0.4.4
distlib==0.3.3
filelock==3.3.0
Flask==2.0.2
Flask-MySQLdb==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.7.0
regex==2021.9.30
six==1.16.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
virtualenv==20.8.1
Werkzeug==2.0.2
